snnreadout moved from utils to contrib
