+++ hls4ml/contrib/snntorch.py
